% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dane_bw.R
\docType{data}
\name{Dane_bw}
\alias{Dane_bw}
\title{Example data}
\format{
\subsection{\code{moj_df}}{

A data frame with 10 observations on 2 variables.
\describe{
\item{x}{numeric vector}
\item{y}{numeric vector}
}
}
}
\source{
Manually created.
}
\usage{
Dane_bw
}
\description{
Example data created manually for the purpose of this package.
}
\keyword{datasets}
